+++ geotext/models/country.py
@@ -1,0 +1,6 @@
+# -*- coding: utf-8 -*-
+from geotext.models.place import Place
+
+
+class Country(Place):
+    pass